+++ WebFlix/WebFlix/domain/Forfait.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Text;
 using System.Collections.Generic;
 using NHibernate.Validator;
 
@@ -9,15 +7,15 @@ namespace WebFlix.Domain {
         public Forfait() {
 			Client = new List<Client>();
         }
-        public virtual decimal Forfaitid { get; set; }
+        public virtual int Forfaitid { get; set; }
         [NotNullNotEmpty]
         [Length(13)]
         public virtual string Nomforfait { get; set; }
         public virtual double? Cout { get; set; }
         [NotNullNotEmpty]
-        public virtual decimal Locationmax { get; set; }
+        public virtual int Locationmax { get; set; }
         [NotNullNotEmpty]
-        public virtual decimal Dureemax { get; set; }
+        public virtual int Dureemax { get; set; }
         public virtual IList<Client> Client { get; set; }
     }
 }
