+++ WebFlix/WebFlix/domain/Genre.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Text;
 using System.Collections.Generic;
 using NHibernate.Validator;
 
@@ -8,7 +6,7 @@ namespace WebFlix.Domain {
         public Genre() {
 			Filmgenre = new List<Filmgenre>();
         }
-        public virtual decimal Genreid { get; set; }
+        public virtual int Genreid { get; set; }
         [NotNullNotEmpty]
         [Length(50)]
         public virtual string Nomgenre { get; set; }
