+++ WebFlix/WebFlix/domain/Informationpersonel.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Text;
 using System.Collections.Generic;
 using NHibernate.Validator;
 
@@ -11,7 +10,7 @@ namespace WebFlix.Domain {
 			Acteur = new List<Acteur>();
 			Scenariste = new List<Scenariste>();
         }
-        public virtual decimal Personelid { get; set; }
+        public virtual int Personelid { get; set; }
         [Length(50)]
         public virtual string Nom { get; set; }
         [Length(50)]
