+++ WebFlix/WebFlix/domain/Location.cs
@@ -1,12 +1,10 @@
 using System;
-using System.Text;
-using System.Collections.Generic;
 using NHibernate.Validator;
 
 namespace WebFlix.Domain {
     
     public class Location {
-        public virtual decimal Locationid { get; set; }
+        public virtual int Locationid { get; set; }
         public virtual Film Film { get; set; }
         public virtual Users Users { get; set; }
         [NotNullNotEmpty]
