+++ WebFlix/WebFlix/domain/Realisateur.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Text;
 using System.Collections.Generic;
 using NHibernate.Validator;
 
@@ -9,7 +7,7 @@ namespace WebFlix.Domain {
         public Realisateur() {
 			Film = new List<Film>();
         }
-        public virtual decimal Realisateurid { get; set; }
+        public virtual int Realisateurid { get; set; }
         public virtual Informationpersonel Informationpersonel { get; set; }
         [Length(100)]
         public virtual string Nomrealisateur { get; set; }
