+++ WebFlix/WebFlix/domain/Role.cs
@@ -1,12 +1,9 @@
-using System;
-using System.Text;
-using System.Collections.Generic;
 using NHibernate.Validator;
 
 namespace WebFlix.Domain {
     
     public class Role {
-        public virtual decimal Roleid { get; set; }
+        public virtual int Roleid { get; set; }
         public virtual Film Film { get; set; }
         public virtual Acteur Acteur { get; set; }
         [Length(100)]
